FEAT Add the Command MODE with the option t,k,o,i and l, all can take only one args, the error return to the client is done. Maybe the way to the serv send error is not the best but its working
